Updated all files to issue an error when called from Python 3. Pyradi google code home page updated to warn not to use Python 3 for pyradi.
All forms of the Planck function approach an infinite value in the exponent when used
with short wavelengths and low temperatures.  The code now checks for these conditions
and handles it by returning a zero exitance values.

Fixed a bug when plotting polar plots with negative values.
Fixed a shape bug in rychroma example.

All files tested and passed with my PC status on Python 2.7.6, Matplotlib 1.3.1, Numpy 1.8.0 and Scipy 0.13.3.