Added calibration functionality to ryptw. If you have the sensor's calibration data. ryptw can now calculate source temperature and radiance. Added a small dataset to test. The calibration data must be in an XML file, format as given in data/lwir100mm150us10ND20090910.xml. The XML file format is reasonably self-explanatory, except the DLFloor and Power: these define the instrument self-radiance knee at low digital levels. Power determines the sharpness of the knee and DLFloor determines the asymptotic level. Both are determined interactively, set these to such values to best match the measured calibration points.
Added a function to ryutls to set up matplotlib to display an upright greek mu.
Upright mu symbols are required to indicate the SI variation 'micro'.
This function slows down execution, so use with care.

Added a function to ryplot to perform histogram equalisation on an image. 
The function allows standard equalisation and also provides for two compression
laws to be be applied prior to equalisation.  These compression laws sometimes
give a better result.

Added the 'marker' parameter to all cartesian plots, to enable data points to be
displayed with matplotlib's standard line marker symbols ('x', 'o', etc.).

Refactored ryplanck to prevent a remaining overflow on long wavelengths and low 
temperatures.  This overflow is inevitable because of the extreme value of x.

Updated pdf and html documentation.